[fix] 푸쉬 알림 함수 분리 및 로직 수정
